Bonus task one: keyboard focus on `<button>`
The lesson here is that if you use the right semantics, you get things
like keyboard-reachable elements for free. I've added a small bit of CSS
that's largely gratuitous.

That being said, the trick here is to make sure your operating system
and browser itself is configured to respond to Tab events. On MacOS, for
example, you have to make sure that All Controls is selected in System
Preferences > Keyboard > Shortcuts.