Bonus task four: visually indicate doorbell ring
It's events all the way down. Here I'm just registering listeners on the
`doorbell` object for when it begins to play, and when its playback has
ended. On those events, I just add a message to the DOM that the
doorbell rings, and remove it as soon as playback is done.

There are other approaches here, mostly with Promise objects that Audio
is supposed to return -- but those are not as widely supported as good
old media events. See the Mozilla docs on HTMLMediaElement and
especially the Events section:

https://developer.mozilla.org/en-US/docs/Web/API/HTMLMediaElement